Generalize the foreign callback storing code
Added a macro to implement foreign callback cells.  These are types that
store pointers to foreign callback functions in an AtomicUsize.  I'm
planning to add a couple in the near future to extend the async
capabilities.

One functional change is that we will no longer panic if a callback is
set multiple times.  The callback interface code did that, but the async
code didn't.  I've never seen the callback interface code actually
generate the error, so I think we can get rid of the check in order to
standarize things.

Put the code for this in `foreigncallbacks` and moved the code that
implements callback interfaces into `callbackinterface.rs`.
