+++ uniffi_core/src/ffi/mod.rs
@@ -4,6 +4,7 @@
 
 //! Types that can cross the FFI boundary.
 
+pub mod callbackinterface;
 pub mod ffidefault;
 pub mod foreignbytes;
 pub mod foreigncallbacks;
@@ -12,6 +13,7 @@ pub mod rustbuffer;
 pub mod rustcalls;
 pub mod rustfuture;
 
+pub use callbackinterface::*;
 pub use ffidefault::FfiDefault;
 pub use foreignbytes::*;
 pub use foreigncallbacks::*;
